Software/Python ProLamp Framework/python-opencv/requirements.txt
numpy>=1.20.0
opencv-python
opencv-contrib-python>=4.5.0.64
pyyaml
black